pyproject.toml
[tool.poetry]
name = "mintedcsv"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
usaddress-scourgify = "^0.2.4"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
ipdb = "^0.13.9"
mypy = "^0.971"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
